テストのパフォーマンスを少し改善
