+++ README.md
@@ -6,6 +6,7 @@ React demo site program
 ## Development(1)
 | Item   | Ver. |Remarks|
 |--------|--------|--------|
+| cpu    | Intel(R) Core(TM) i5-5250U CPU @ 1.60GHz||
 | os     | Darwin 21.6.0 |Monterey|
 | node    | 18.16.0||
 | react   | 18.2.0||
@@ -13,6 +14,7 @@ React demo site program
 ## Development(2)
 | Item   | Ver. |Remarks|
 |--------|--------|--------|
+| cpu     | Intel(R) Core(TM) i7-10510U CPU @ 1.80GHz ||
 | os     | Ubuntu 20.04 LTS |WSL2|
 | node    | 18.16.0||
 | react   | 18.2.0||
