+++ package.json
@@ -21,7 +21,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "jest --env=jsdom",
+    "test": "jest --env=jsdom --maxWorkers=3",
     "lint": "eslint src --ext .js --ext .jsx",
     "lint:fix": "eslint src --fix"
   },
